Add files via upload
The Train-Test split is a crucial step in any machine learning model prediction. I divided my data on an 80-20 basis and trained the models using Random Forest Classifier, XGBoost, and SVM. It turns out that the classifier algorithms provided the highest and most accurate predictions. I compared the performance of these models in the final step.